# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "媒资管理系统 REST API。成功响应直接返回业务 JSON（camelCase）；失败返回 { err_code, err_msg }。",
        "title": "GIN-MAM API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/auth/login": {
            "post": {
                "description": "用户名密码登录，返回 JWT Token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "认证"
                ],
                "summary": "用户登录",
                "parameters": [
                    {
                        "description": "登录参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_domain_sys.loginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_domain_sys.LoginResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_gin-mam_backend_internal_pkg_httpx.ErrorVo"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_gin-mam_backend_internal_pkg_httpx.ErrorVo"
                        }
                    }
                }
            }
        },
        "/api/v1/catalog": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "编目"
                ],
                "summary": "创建编目节点",
                "parameters": [
                    {
                        "description": "编目参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_domain_catalog.createRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/internal_domain_catalog.CatalogVO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_gin-mam_backend_internal_pkg_httpx.ErrorVo"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_gin-mam_backend_internal_pkg_httpx.ErrorVo"
                        }
                    }
                }
            }
        },
        "/api/v1/catalog/tree": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "返回完整多级编目树结构",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "编目"
                ],
                "summary": "获取编目树",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/internal_domain_catalog.TreeNode"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_gin-mam_backend_internal_pkg_httpx.ErrorVo"
                        }
                    }
                }
            }
        },
        "/api/v1/catalog/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "编目"
                ],
                "summary": "更新编目节点",
                "parameters": [
                    {
                        "type": "string",
                        "description": "编目 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "更新参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_domain_catalog.updateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_domain_catalog.CatalogVO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_gin-mam_backend_internal_pkg_httpx.ErrorVo"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_gin-mam_backend_internal_pkg_httpx.ErrorVo"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_gin-mam_backend_internal_pkg_httpx.ErrorVo"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "编目"
                ],
                "summary": "删除编目节点",
                "parameters": [
                    {
                        "type": "string",
                        "description": "编目 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_gin-mam_backend_internal_pkg_httpx.ErrorVo"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_gin-mam_backend_internal_pkg_httpx.ErrorVo"
                        }
                    }
                }
            }
        },
        "/api/v1/catalog/{id}/config": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "编目"
                ],
                "summary": "列出编目配置",
                "parameters": [
                    {
                        "type": "string",
                        "description": "编目 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/internal_domain_catalog.ConfigVO"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_gin-mam_backend_internal_pkg_httpx.ErrorVo"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_gin-mam_backend_internal_pkg_httpx.ErrorVo"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "编目"
                ],
                "summary": "批量更新编目配置",
                "parameters": [
                    {
                        "type": "string",
                        "description": "编目 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "配置项列表",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_domain_catalog.batchConfigRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/internal_domain_catalog.ConfigVO"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_gin-mam_backend_internal_pkg_httpx.ErrorVo"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_gin-mam_backend_internal_pkg_httpx.ErrorVo"
                        }
                    }
                }
            }
        },
        "/api/v1/search/asset": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "基于 Elasticsearch 全文检索媒资，命中后回源拼装 AssetVO",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "检索"
                ],
                "summary": "检索媒资",
                "parameters": [
                    {
                        "description": "检索参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_domain_search.searchAssetRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_gin-mam_backend_internal_domain_asset.PageResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_gin-mam_backend_internal_pkg_httpx.ErrorVo"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_gin-mam_backend_internal_pkg_httpx.ErrorVo"
                        }
                    }
                }
            }
        },
        "/api/v1/sys/user/page": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取系统用户分页列表",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "系统用户"
                ],
                "summary": "分页查询用户",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "每页条数",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_domain_sys.UserPageResult"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_gin-mam_backend_internal_pkg_httpx.ErrorVo"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_gin-mam_backend_internal_domain_asset.AssetVO": {
            "type": "object",
            "properties": {
                "catalogId": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "createdBy": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "metadata": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "previewUrl": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "storagePath": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "github_com_gin-mam_backend_internal_domain_asset.PageResult": {
            "type": "object",
            "properties": {
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_gin-mam_backend_internal_domain_asset.AssetVO"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "github_com_gin-mam_backend_internal_pkg_httpx.ErrorVo": {
            "type": "object",
            "properties": {
                "err_code": {
                    "type": "integer",
                    "example": 40100
                },
                "err_msg": {
                    "type": "string",
                    "example": "未授权"
                }
            }
        },
...
        "internal_domain_catalog.CatalogVO": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "parentId": {
                    "type": "string"
                },
                "sortCode": {
                    "type": "integer"
                }
            }
        },
        "internal_domain_catalog.ConfigVO": {
            "type": "object",
            "properties": {
                "configKey": {
                    "type": "string"
                },
                "configValue": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                }
            }
        },
        "internal_domain_catalog.TreeNode": {
            "type": "object",
            "properties": {
                "children": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_domain_catalog.TreeNode"
                    }
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "parentId": {
                    "type": "string"
                }
            }
        },
        "internal_domain_catalog.batchConfigRequest": {
            "type": "object",
            "required": [
                "items"
            ],
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "required": [
                            "configKey"
                        ],
                        "properties": {
                            "configKey": {
                                "type": "string"
                            },
                            "configValue": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "internal_domain_catalog.createRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string",
                    "example": "新闻栏目"
                },
                "parentId": {
                    "type": "string",
                    "example": ""
                },
                "sortCode": {
                    "type": "integer",
                    "example": 0
                }
            }
        },
        "internal_domain_catalog.updateRequest": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "parentId": {
                    "type": "string"
                },
                "sortCode": {
                    "type": "integer"
                }
            }
        },
        "internal_domain_search.searchAssetRequest": {
            "type": "object",
            "properties": {
                "catalogId": {
                    "type": "string",
                    "example": ""
                },
                "keyword": {
                    "type": "string",
                    "example": "宣传片"
                },
                "page": {
                    "type": "integer",
                    "example": 1
                },
                "pageSize": {
                    "type": "integer",
                    "example": 20
                },
                "type": {
                    "type": "string",
                    "example": "video"
                }
            }
        },
        "internal_domain_sys.LoginResult": {
            "type": "object",
            "properties": {
                "expiresAt": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                }
            }
        },
        "internal_domain_sys.UserPageResult": {
            "type": "object",
            "properties": {
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_domain_sys.UserVO"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "internal_domain_sys.UserVO": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "nickname": {
                    "type": "string"
                },
                "orgId": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "internal_domain_sys.loginRequest": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "type": "string",
                    "example": "admin123"
                },
                "username": {
                    "type": "string",
                    "example": "admin"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT Bearer Token，格式：Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}